move common layout package
