+++ app/build.gradle
@@ -56,6 +56,7 @@ dependencies {
     implementation project(":data:repository")
     implementation project(":common:util")
     implementation project(":common:navigation")
+    implementation project(":common:layout")
     implementation project(":feature:list_database")
     implementation project(":feature:list_qiita_doc")
     implementation project(":feature:select")
