+++ feature/list_qiita_doc/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/feature/list_qiita_doc/ListQiitaDocFragment.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.afri_inc.senri.utils.autoCleared
 import dagger.android.support.DaggerFragment
-import jp.chau2chaun2.honkot.sample.multimodule.feature.list_qiita_doc.databinding.FragmentListBinding
+import jp.chau2chaun2.honkot.sample.multimodule.common.laout.databinding.FragmentListBinding
 import jp.chau2chaun2.honkot.sample.repository.ImageRepository
 import javax.inject.Inject
